service/backup: split manifest load and deletion
Loading all manifests in parallel can be memory-intensive. This splits
the loading of manifests and the actual file deletion in the purge step.
This way by default manifests are loaded only one at a time, but a
configurable parallelism can be provided if memory is not as constrained
and they can be loaded in parallel again.

fixes #3113